Change type of data int8_t

+++ ESP32_L298N.cpp
@@ -1,7 +1,7 @@
 #include "ESP32_L298N.h"
 
 
-void L298N::moveMotor(int8_t pwmA, int8_t pwmB){
+void L298N::moveMotor(int pwmA, int pwmB){
   if (dl298n.pin3 != dl298n.pin4){
     analogWrite(dl298n.pinA,pwmA);
     analogWrite(dl298n.pinB,pwmB);
@@ -15,7 +15,7 @@ void L298N::moveMotor(int8_t pwmA, int8_t pwmB){
     digitalWrite(dl298n.pin2,0);
   }  
 }
-void L298N::moveMotor(int8_t pwmA){
+void L298N::moveMotor(int pwmA){
   if (dl298n.pin3 != dl298n.pin4){
     analogWrite(dl298n.pinA, pwmA);
     analogWrite(dl298n.pinB, pwmA);
@@ -29,7 +29,7 @@ void L298N::moveMotor(int8_t pwmA){
     digitalWrite(dl298n.pin2,0);
   }    
 }
-void L298N::backMotor(int8_t pwmA, int8_t pwmB){
+void L298N::backMotor(int pwmA, int pwmB){
   if (dl298n.pin3 != dl298n.pin4){
     analogWrite(dl298n.pinA,pwmA);
     analogWrite(dl298n.pinB,pwmB);
@@ -43,7 +43,7 @@ void L298N::backMotor(int8_t pwmA, int8_t pwmB){
     digitalWrite(dl298n.pin2,1);
   }
 }
-void L298N::backMotor(int8_t pwmA){
+void L298N::backMotor(int pwmA){
   if (dl298n.pin3 != dl298n.pin4){
     analogWrite(dl298n.pinA, pwmA);
     analogWrite(dl298n.pinB, pwmA);
@@ -58,17 +58,17 @@ void L298N::backMotor(int8_t pwmA){
   }
         
 }
-void L298N::moveMotor12(int8_t pwmA){
+void L298N::moveMotor12(int pwmA){
   analogWrite(dl298n.pinA, pwmA);
   digitalWrite(dl298n.pin1,1);
   digitalWrite(dl298n.pin2,0);
 }
-void L298N::backMotor12(int8_t pwmA){
+void L298N::backMotor12(int pwmA){
   analogWrite(dl298n.pinA, pwmA);
   digitalWrite(dl298n.pin1,0);
   digitalWrite(dl298n.pin2,1);
 }
-void L298N::moveMotor34(int8_t pwmA){
+void L298N::moveMotor34(int pwmA){
   if (dl298n.pin3 != dl298n.pin4){
     analogWrite(dl298n.pinB, pwmA);
     digitalWrite(dl298n.pin3,1);
@@ -79,7 +79,7 @@ void L298N::moveMotor34(int8_t pwmA){
     digitalWrite(dl298n.pin2,0);
   }
 }
-void L298N::backMotor34(int8_t pwmA){
+void L298N::backMotor34(int pwmA){
   if (dl298n.pin3 != dl298n.pin4){
     analogWrite(dl298n.pinB, pwmA);
     digitalWrite(dl298n.pin3,0);
